feat: Complete TODOs and get project running
This commit completes all of the TODOs for the project rubric. The code has been tested in the simulation and performs well. There is one brief period where the predicted location swings away from the car but it quickly comes back. It is unclear what causes that, but it shows the ability for the Kalman filter to quickly correct for noise in the system.